creating selector for question on test

+++ src/main/java/com/exam/license/exam/models/Question.java
@@ -45,8 +45,8 @@ public class Question {
         return id;
     }
 
-    public void setId(Long id) {
-        this.id = id;
+    public void setId(int id) {
+        this.id = Long.valueOf(id);
     }
 
     public String getQuestion() {
